refactor(examples): create one single example binary instead of multiple equal ones
