+++ vibe/Cargo.toml
@@ -9,14 +9,15 @@ repository = "https://github.com/TornaxO7/vibe"
 publish = false
 
 [dependencies]
+anyhow.workspace = true
 serde.workspace = true
 shady-audio.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 wgpu.workspace = true
 cgmath.workspace = true
+clap.workspace = true
 
-anyhow = "1"
 raw-window-handle = "0.6"
 
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
@@ -31,4 +32,3 @@ toml = "0.8"
 
 winit = "0.30"
 notify = "8"
-clap = { version = "4.5", features = ["derive"] }
